add button validation code

+++ Project/Bahan/PenerimaanKain.cs
@@ -17,21 +17,36 @@ namespace Project
         public PenerimaanKain()
         {
             InitializeComponent();
+            cboNoPOKain.Text = "Select PO Number";
         }
 
         private void btnExitPenerimaanKain_Click(object sender, EventArgs e)
         {
-            Close();
+            cboNoPOKain.Text = "";
+            txtSupplierCode.Clear();
+            txtSupplierName.Clear();
+            this.Close();
         }
 
         private void cboNoPOKain_SelectedIndexChanged(object sender, EventArgs e)
         {
-            long poNumber = Convert.ToInt64(cboNoPOKain.SelectedValue.ToString());
-            var db = GenericQuery.SqlQuerySingle<POKainModel>("SELECT p.idPOKain, p.PONumber, p.SupplierID, s.SupplierName, s.SupplierCode FROM PreOrderKains p JOIN IndomodaSuppliers s ON p.SupplierID = s.SupplierID WHERE p.PONumber = '" + poNumber + "'");
-            txtSupplierCode.Text = db.SupplierCode.ToString();
-            txtSupplierName.Text = db.SupplierName.ToString();
-            List<DetailPOModel> listPO = GenericQuery.SqlQuery<DetailPOModel>("SELECT * FROM DetailPO WHERE PONumber = '" + poNumber + "'");
-            detailPOBindingSource.DataSource = listPO.ToList();
+            try
+            {
+                if (cboNoPOKain.Items.Count > 0 && cboNoPOKain.Text != "")
+                {
+                    cboNoPOKain.Refresh();
+                    long poNumber = Convert.ToInt64(cboNoPOKain.SelectedValue.ToString());
+                    var db = GenericQuery.SqlQuerySingle<POKainModel>("SELECT p.idPOKain, p.PONumber, p.SupplierID, s.SupplierName, s.SupplierCode FROM PreOrderKains p JOIN IndomodaSuppliers s ON p.SupplierID = s.SupplierID WHERE p.PONumber = '" + poNumber + "'");
+                    txtSupplierCode.Text = db.SupplierCode.ToString();
+                    txtSupplierName.Text = db.SupplierName.ToString();
+                    List<DetailPOModel> listPO = GenericQuery.SqlQuery<DetailPOModel>("SELECT * FROM DetailPO WHERE PONumber = '" + poNumber + "'");
+                    detailPOBindingSource.DataSource = listPO.ToList();
+                }
+            }
+            catch (NullReferenceException ex)
+            {
+                MetroFramework.MetroMessageBox.Show(this, ex.Message, "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
         }
 
         private void PenerimaanKain_Load(object sender, EventArgs e)
@@ -43,28 +58,39 @@ namespace Project
                 preOrderKainBindingSource.DataSource = listCboPO.ToList();
                 materialBindingSource.DataSource = db.Materials.ToList();
                 colorBindingSource.DataSource = db.Colors.ToList();
+                cboNoPOKain.Refresh();
             }
         }
 
         private void btnDeletePenerimaanKain_Click(object sender, EventArgs e)
         {
-            if (MetroFramework.MetroMessageBox.Show(this, "Do you want to delete this PO?", "Message", MessageBoxButtons.YesNo, MessageBoxIcon.Error) == DialogResult.Yes)
+            if (cboNoPOKain.Items.Count == 0)
             {
-                using (indomodaEntities db = new indomodaEntities())
+                MetroFramework.MetroMessageBox.Show(this, "PO List is empty!", "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else
+            {
+                if (MetroFramework.MetroMessageBox.Show(this, "Do you want to delete this PO?", "Message", MessageBoxButtons.YesNo, MessageBoxIcon.Error) == DialogResult.Yes)
                 {
-                    try
+                    using (indomodaEntities db = new indomodaEntities())
                     {
-                        long poNumber = Convert.ToInt64(cboNoPOKain.SelectedValue.ToString());
-                        int b = GenericQuery.ExecSQLCommand("DELETE FROM PreOrderKains WHERE poNumber = '" + poNumber + "'");
-                        db.SaveChangesAsync().Wait();
-                        cboNoPOKain.Refresh();
-                        preOrderKainBindingSource.DataSource = db.PreOrderKains.ToList();
+                        try
+                        {
+                            long poNumber = Convert.ToInt64(cboNoPOKain.SelectedValue.ToString());
+                            int b = GenericQuery.ExecSQLCommand("DELETE FROM PreOrderKains WHERE poNumber = '" + poNumber + "'");
+                            db.SaveChangesAsync().Wait();
+                            cboNoPOKain.Refresh();
+                            dataGridView1.Refresh();
+                            List<PreOrderKain> listCboPO = GenericQuery.SqlQuery<PreOrderKain>("SELECT p.idPOKain, p.PONumber, p.SupplierID, p.status, p.SupplierID, p.GrandTotal, p.Date_time FROM PreOrderKains p WHERE p.status = '" + 0 + "'");
+                            listCboPO.Remove(new PreOrderKain() { PONumber = poNumber });
+                            preOrderKainBindingSource.DataSource = listCboPO.ToList();
 
-                        MetroFramework.MetroMessageBox.Show(this, "Success! This PO has been deleted to the database", "Message", MessageBoxButtons.OK, MessageBoxIcon.Question);
-                    }
-                    catch (Exception ex)
-                    {
-                        MetroFramework.MetroMessageBox.Show(this, ex.Message, "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                            MetroFramework.MetroMessageBox.Show(this, "Success! This PO has been deleted from the database", "Message", MessageBoxButtons.OK, MessageBoxIcon.Question);
+                        }
+                        catch (Exception ex)
+                        {
+                            MetroFramework.MetroMessageBox.Show(this, ex.Message, "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        }
                     }
                 }
             }
@@ -72,26 +98,33 @@ namespace Project
 
         private void btnUpdateStatus_Click(object sender, EventArgs e)
         {
-            if (MetroFramework.MetroMessageBox.Show(this, "Do you want to update this detail PO status?", "Message", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            if (dataGridView1.Rows.Count == 0)
             {
-                using (indomodaEntities db = new indomodaEntities())
+                MetroFramework.MetroMessageBox.Show(this, "Detail PO is empty!", "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else
+            {
+                if (MetroFramework.MetroMessageBox.Show(this, "Do you want to update this detail PO status?", "Message", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
                 {
-                    try
+                    using (indomodaEntities db = new indomodaEntities())
                     {
-                        int currentRow = dataGridView1.CurrentRow.Index;
-                        int currentDetailPOID = Convert.ToInt32(dataGridView1[0, dataGridView1.CurrentRow.Index].Value.ToString());
-                        bool setStatus = Convert.ToBoolean(dataGridView1.Rows[currentRow].Cells[6].Value);
-                        int a = GenericQuery.ExecSQLCommand("UPDATE DetailPO SET DetailStatus = @DetailStatus WHERE DetailPOID = '" + currentDetailPOID + "'", new[] {
+                        try
+                        {
+                            int currentRow = dataGridView1.CurrentRow.Index;
+                            int currentDetailPOID = Convert.ToInt32(dataGridView1[0, dataGridView1.CurrentRow.Index].Value.ToString());
+                            bool setStatus = Convert.ToBoolean(dataGridView1.Rows[currentRow].Cells[6].Value);
+                            int a = GenericQuery.ExecSQLCommand("UPDATE DetailPO SET DetailStatus = @DetailStatus WHERE DetailPOID = '" + currentDetailPOID + "'", new[] {
                             new SqlParameter("@DetailStatus", setStatus)
                         });
-                        db.SaveChangesAsync().Wait();
-                        dataGridView1.Refresh();
+                            db.SaveChangesAsync().Wait();
+                            dataGridView1.Refresh();
 
-                        MetroFramework.MetroMessageBox.Show(this, "Success! This detail PO stauts has been marked as done", "Message", MessageBoxButtons.OK, MessageBoxIcon.Question);
-                    }
-                    catch (Exception ex)
-                    {
-                        MetroFramework.MetroMessageBox.Show(this, ex.Message, "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                            MetroFramework.MetroMessageBox.Show(this, "Success! This detail PO stauts has been updated", "Message", MessageBoxButtons.OK, MessageBoxIcon.Question);
+                        }
+                        catch (Exception ex)
+                        {
+                            MetroFramework.MetroMessageBox.Show(this, ex.Message, "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        }
                     }
                 }
             }
@@ -99,29 +132,63 @@ namespace Project
 
         private void btnSaveUpdateStatus_Click(object sender, EventArgs e)
         {
-            long poNumber = Convert.ToInt64(cboNoPOKain.SelectedValue.ToString());
-            List<DetailPOModel> listPO = GenericQuery.SqlQuery<DetailPOModel>("SELECT * FROM DetailPO WHERE PONumber = '" + poNumber + "' AND DetailStatus = '"+0+"'");
-            if (listPO.Count > 0)
+            if (cboNoPOKain.Items.Count == 0)
             {
-                MetroFramework.MetroMessageBox.Show(this, "You must mark all the Detail PO as done first!", "Message", MessageBoxButtons.OK, MessageBoxIcon.Question);
+                MetroFramework.MetroMessageBox.Show(this, "PO List is empty!", "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else if (cboNoPOKain.Items.Count > 0 && txtSupplierCode.Text == "" && txtSupplierName.Text == "")
+            {
+                MetroFramework.MetroMessageBox.Show(this, "You must select the PO Number first!", "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else if (cboNoPOKain.Items.Count > 0 && txtSupplierCode.Text != "" && txtSupplierName.Text != "" && dataGridView1.Rows.Count == 0)
+            {
+                MetroFramework.MetroMessageBox.Show(this, "You must select the PO Number first!", "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             else
             {
-                using (indomodaEntities db = new indomodaEntities())
+                long po = Convert.ToInt64(cboNoPOKain.SelectedValue.ToString());
+                List<DetailPOModel> listPODb = GenericQuery.SqlQuery<DetailPOModel>("SELECT * FROM DetailPO WHERE PONumber = '" + po + "' AND DetailStatus = '" + 0 + "'");
+
+                if (listPODb.Count > 0)
                 {
-                    try
+                    MetroFramework.MetroMessageBox.Show(this, "You must mark all the Detail PO as done first!", "Message", MessageBoxButtons.OK, MessageBoxIcon.Question);
+                }
+                else
+                {
+                    using (indomodaEntities db = new indomodaEntities())
                     {
-                        int setStatus = 1;
-                        int a = GenericQuery.ExecSQLCommand("UPDATE PreOrderKains SET status = @status WHERE PONumber = '" + poNumber + "'", new[] {          
+                        try
+                        {
+                            int setStatus = 1;
+                            int a = GenericQuery.ExecSQLCommand("UPDATE PreOrderKains SET status = @status WHERE PONumber = '" + po + "'", new[] {
                                 new SqlParameter("@status", setStatus)
                             });
-                        db.SaveChangesAsync().Wait();
+                            db.SaveChangesAsync().Wait();
+                            cboNoPOKain.Refresh();
+
+                            if (cboNoPOKain.Items.Count == 0)
+                            {
+                                dataGridView1.Rows.Clear();
+                                dataGridView1.Refresh();
+                            }
+                            else
+                            {
+                                dataGridView1.Rows.Clear();
+                            }
+
+                            btnSaveUpdateStatus.Refresh();
+                            List<PreOrderKain> listCboPO = GenericQuery.SqlQuery<PreOrderKain>("SELECT p.idPOKain, p.PONumber, p.SupplierID, p.status, p.SupplierID, p.GrandTotal, p.Date_time FROM PreOrderKains p WHERE p.status = '" + 0 + "'");
+                            listCboPO.Remove(new PreOrderKain() { PONumber = po });
+                            preOrderKainBindingSource.DataSource = listCboPO.ToList();
+                            txtSupplierCode.Clear();
+                            txtSupplierName.Clear();
+                            MetroFramework.MetroMessageBox.Show(this, "Success! This PO has been marked as done", "Message", MessageBoxButtons.OK, MessageBoxIcon.Question);
+                        }
+                        catch (Exception ex)
+                        {
+                            MetroFramework.MetroMessageBox.Show(this, ex.Message, "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        }
                     }
-                    catch (Exception ex)
-                    {
-                        MetroFramework.MetroMessageBox.Show(this, ex.Message, "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                    }
-                    this.Close();
                 }
             }
         }
